feat. pwdreset etc

+++ frontend/src/stores/users/auth.store.js
@@ -1,11 +1,9 @@
 import { defineStore } from 'pinia'
-
 import { fetchWrapper } from '@/helpers'
 import router from '@/router'
 import { useAlertStore } from '@/stores'
 
 const baseUrl = `${import.meta.env.VITE_API_URL}/users`
-
 
 export const useAuthStore = defineStore({
   id: 'auth',
@@ -15,24 +13,48 @@ export const useAuthStore = defineStore({
     returnUrl: null,
   }),
   actions: {
-    async login(username, password) {
+    async login(email, password, keeplog) {
       try {
-        console.log('로그인 함수 도착')
+        // baseUrl/users/login/{requestBody}
         const user = await fetchWrapper.post(`${baseUrl}/authenticate`, {
-          username,
+          email,
           password,
+          keeplog,
         })
-        console.log(user);
-
+        console.log(user)
 
         // update pinia state
         this.user = user
 
-        // store user details and jwt in local storage to keep user logged in between page refreshes
-        localStorage.setItem('user', JSON.stringify(user))
-
-        // redirect to previous url or default to home page
+        // 페이지 새로고침을해도 유저가 로그인을 유지하도록 유저 디테일과 jwt를 로컬스토리지에 저장 -> 일단은 로그인유지를 클릭시에만 가능
+        if (keeplog === true) {
+          localStorage.setItem('user', JSON.stringify(user))
+        }
+        // 기본 페이지로 리다이렉트
         router.push(this.returnUrl || '/')
+      } catch (error) {
+        const alertStore = useAlertStore()
+        alertStore.error(error)
+      }
+    },
+    async passwordreset(username, email) {
+      try {
+        console.log('>>>>암호 리셋함수 도착<<<<')
+        // baseUrl/users/pwdreset/{requestBody}
+      } catch (error) {
+        const alertStore = useAlertStore()
+        alertStore.error(error)
+      }
+      router.push(this.returnUrl || '/')
+    },
+    async deactivate() {
+      try {
+        console.log('>>>>비활성화 함수 도착<<<<')
+        console.log(this.user)
+        // baseUrl/users/{id}
+        this.user = null
+        localStorage.removeItem('user')
+        router.push('/about')
       } catch (error) {
         const alertStore = useAlertStore()
         alertStore.error(error)
